Rename VoiceTest into VoiceDemo
.. to make more ovbious that it's not a unit test
